* Upgrade presets for DNNL 3.5.3, OpenBLAS 0.3.28, CPython 3.12.5, NumPy 2.0.1, LLVM 18.1.8, OpenCL 3.0.16, ONNX 1.16.2, ONNX Runtime 1.18.1, TVM 0.17.0

+++ dnnl/samples/CpuCnnInferenceInt8.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2018-2022 Intel Corporation
+* Copyright 2018-2024 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -87,7 +87,7 @@ public class CpuCnnInferenceInt8 {
 
         //[Choose scaling factors]
 
-        /// The *source, weights, bias* and *destination* datasets use the single-scale
+        /// The *source, weights* and *destination* datasets use the single-scale
         /// format with mask set to '0'.
         /// @snippet cnn_inference_int8.cpp Set scaling mask
         //[Set scaling mask]
@@ -122,7 +122,7 @@ public class CpuCnnInferenceInt8 {
         /// descriptors are configured as:
         ///
         /// * 8-bit unsigned (u8) for source and destination.
-        /// * 8-bit signed (s8) for bias and weights.
+        /// * 8-bit signed (s8) for weights.
         ///
         ///  > **Note**
         ///  > The destination type is chosen as *unsigned* because the
@@ -130,7 +130,7 @@ public class CpuCnnInferenceInt8 {
         /// @snippet cnn_inference_int8.cpp Create convolution memory descriptors
         //[Create convolution memory descriptors]
         memory.desc conv_src_md = new memory.desc(conv_src_tz, memory.data_type.u8, memory.format_tag.any);
-        memory.desc conv_bias_md = new memory.desc(conv_bias_tz, memory.data_type.s8, memory.format_tag.any);
+        memory.desc conv_bias_md = new memory.desc(conv_bias_tz, memory.data_type.f32, memory.format_tag.any);
         memory.desc conv_weights_md = new memory.desc(conv_weights_tz, memory.data_type.s8, memory.format_tag.any);
         memory.desc conv_dst_md = new memory.desc(conv_dst_tz, memory.data_type.u8, memory.format_tag.any);
         //[Create convolution memory descriptors]
